Separate production and test code
- Define sibling unit-test suites and routers explicitly.
- Keep test-only helpers and fixtures in test-scoped files and modules.
- Align coverage filters with the test-file convention.
