+++ crates/ag-tui-text/src/style.rs
@@ -180,18 +180,5 @@ fn active_settings() -> TextRenderSettings {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_default_palette_uses_current_theme_table_surface() {
-        // Arrange
-        let palette = TextPalette::DEFAULT;
-
-        // Act
-        let surface_elevated = palette.surface_elevated;
-
-        // Assert
-        assert_eq!(surface_elevated, Color::Black);
-    }
-}
+#[path = "style_test.rs"]
+mod tests;
